fix: move shared file utils to `/utils/file.ts`
+ some minor cleanup
